chore: moved toastOptions to constants
