+++ src/hooks/useRemoveFromList.tsx
@@ -1,7 +1,7 @@
 import { useSupabaseClient } from '@supabase/auth-helpers-react'
 import { toast } from 'react-toastify'
 
-import { toastOptions } from '@/assets/utilities'
+import { toastOptions } from '@constants'
 
 const useRemoveFromList = (id: string | number, mediaType: string) => {
   const supabase = useSupabaseClient()
@@ -26,19 +26,19 @@ const useRemoveFromList = (id: string | number, mediaType: string) => {
           render: `${type} not in your list`,
           type: "info",
           ...toastOptions
-        });
+        })
       } else if (status === 204) {
         toast.update(toastId, {
           render: `Removed ${type} from your list`,
           type: "success",
           ...toastOptions
-        });
+        })
       } else if (status === 0) {
         toast.update(toastId, {
           render: 'Failed to upload',
           type: "error",
           ...toastOptions
-        });
+        })
       }
     }
   }
